refactor module into aws-parameters layer and restore constants
