+++ gfe-db/pipeline/lambda_layers/aws_parameters/awsparameters/__init__.py
@@ -1,0 +1,1 @@
+from .manager import AppConfig
