+++ gfe-db/pipeline/lambda_layers/aws_parameters/requirements.txt
@@ -1,0 +1,1 @@
+boto3
